+++ scanpipe/pipes/fetch.py
@@ -34,6 +34,7 @@ from urllib.parse import urlparse
 
 from django.conf import settings
 
+import git
 import requests
 from commoncode import command
 from commoncode.hash import multi_checksums
@@ -312,6 +313,28 @@ def fetch_docker_image(docker_url, to=None):
     )
 
 
+def fetch_git_repo(url, to=None):
+    """Fetch provided git ``url`` as a clone and return a ``Download`` object."""
+    download_directory = to or tempfile.mkdtemp()
+    url = url.rstrip("/")
+    filename = url.split("/")[-1]
+    to_path = Path(download_directory) / filename
+    # Disable any prompt, especially for credentials
+    git_env = {"GIT_TERMINAL_PROMPT": "0"}
+
+    git.Repo.clone_from(url=url, to_path=to_path, depth=1, env=git_env)
+
+    return Download(
+        uri=url,
+        directory=download_directory,
+        filename=filename,
+        path=to_path,
+        size="",
+        sha1="",
+        md5="",
+    )
+
+
 SCHEME_TO_FETCHER_MAPPING = {
     "http": fetch_http,
     "https": fetch_http,
@@ -321,6 +344,12 @@ SCHEME_TO_FETCHER_MAPPING = {
 
 def get_fetcher(url):
     """Return the fetcher function based on the provided `url` scheme."""
+    if url.startswith("git@"):
+        raise ValueError("SSH 'git@' URLs are not supported. Use https:// instead.")
+
+    if url.rstrip("/").endswith(".git"):
+        return fetch_git_repo
+
     # Not using `urlparse(url).scheme` for the scheme as it converts to lower case.
     scheme = url.split("://")[0]